feat: 강아지 만 나이 계산 함수 추가
